parse and pass-through arch specific breakpoint kinds
So yeah, turns out that the breakpoint kind is actually pretty
important, especially when working with `gdbstub` in a non-emulated
context where you can't just sniff the PC. In those cases, you often do
actually need to patch the instruction / set an appropriately sized
hardware breakpoint, which you can only do in one of two ways:
- the "kind" field passed by GDB
- knowing the _precise size_ of instruction currently being executed.
Guess which approach is easier?

As expected, this is a breaking change, but hey, if we're already
breaking things, lets break 'em a bit more!

Moreover, I intend to add support for conditional breakpoints soon, and
might need to modify the add/remove_{sw/hw}_breakpoint methods even
more...

+++ src/protocol/commands/_z.rs
@@ -4,7 +4,7 @@ use super::prelude::*;
 pub struct z<'a> {
     pub type_: u8,
     pub addr: &'a [u8],
-    pub kind: u8,
+    pub kind: usize,
 }
 
 impl<'a> ParseCommand<'a> for z<'a> {
